Study-01/web_version/requirements.txt
Pillow==10.0.0
numpy==1.24.3
